Adding database
Add database connection and schema for Termini project

Updated appsettings files to include a new ConnectionStrings section for SQL Server. Added a new project, TerminiDatabase, to the solution for managing database schema and deployment. Created a SQL project file and a schema comparison file for database management. Defined a new SQL schema, Termini, and added a Player table with necessary columns and constraints.